Added package info and restructured dir to for bette dev experience
