+++ src/arg_parser.py
@@ -1,22 +1,28 @@
 
 import os
-from count import count_folders_dir
-from utils import Utils
-from colors import Colors
+import count
+import help_menu
+import utils.colors as color
+
+# FIXME: modules in the same path do not need  from statement
 from rename import Rename
 from spell_checker import SpellChecker
-from config import DrtConfig
-# Interprest cmd string argument
-# and execute cmd
+# from config import DrtConfig
+import utils.config as config
 
+# TODO: Determine the use case and see if you can build on top of this
 OPTIONS = {"-help", "-h", "-V", "-c",
            "-version", "-pr", "-r"}
 COMMANDS = {"config", "check"}
 
 # input format goal drt option or drt cmd arg
 DEFAULT_EDIT_DIR = "{0}/Edit".format(os.path.dirname(os.getcwd()))
+# FIXME: make sure using class desgin properly
 sc = SpellChecker()
-drt = DrtConfig(edit_directory=DEFAULT_EDIT_DIR)
+drt = config.DrtConfig(edit_directory=DEFAULT_EDIT_DIR)
+Colors = color.Colors()
+
+Utils = help_menu.Utils()
 
 
 class ArgParser:
@@ -37,7 +43,7 @@ class ArgParser:
         if(option == "-help" or option == "-h"):
             Utils.display_help()
         elif(option == "-c"):
-            count_folders_dir()
+            count.count_folders_dir()
         elif(option == "-pr"):
             Rename.preview_rename()
         elif(option == "-r"):
