+++ src/rename.py
@@ -11,14 +11,18 @@
 import re
 import os
 import sys
-from colors import Colors
-from global_ import EDITING_DIR
+import utils.colors as color
+import utils.global_ as globals
 
 NAME_PATTERN = re.compile(
     "[a-zA-Z]* ?-?[a-zA-Z]* ?-?[\s]?[a-zA-Z]*[#][(\d+(?:\.\d+)?)]*[-]?.[a-zA-Z]*")
 remove_underscore_dash = "_*-*"
 
+EDITING_DIR = globals.EDITING_DIR
+Colors = color.Colors()
 
+
+# TODO: FIXME: ENSURE PROPER class design -> every class function should have a self parameter
 class Rename:
 
     def rename():
@@ -49,6 +53,7 @@ class Rename:
 
 # TODO: Verify this accepter funciton is no longer needed
 #       might be wise to keep for local testing
+#       TODO: See if there are testing libs for python => possible pytest
 
     def accepter():
